+++ R/service_record.R
@@ -7,8 +7,10 @@
 #' @param filter Filter. One of `'matchmade:pvp'`, `'matchmade:social'`,
 #'   `'matchmade:ranked'`, `'matchmade:bots'`, or `'custom'`.
 #'
-#' @return A `list` with either class `stats_service_record_campaign` or
-#'   `stats_service_record_multiplayer`.
+#' @return For a single gamertag, a `list` with class `campaign` or
+#'   `multiplayer`. For multiple gamertags, a `list` with class `collection`
+#'   where each item in the `list` is a `list` with class `campaign` or
+#'   `multiplayer`.
 #'
 #' @export
 halo_campaign <- function(
@@ -18,19 +20,22 @@ halo_campaign <- function(
     .progress = TRUE
 ) {
   gamertag <- verify_vector_chr(gamertag, 'gamertag')
-  furrr::future_map(
+  resps <- furrr::future_map(
     .x = gamertag,
     .f = function(gamertag) {
       HaloDotAPI(
         endpoint = 'stats/service-record/campaign',
         gamertag = gamertag,
         version = version,
-        token = token
+        token = token,
+        class = 'campaign'
       )
     },
     .options = furrr::furrr_options(seed = NULL),
     .progress = .progress
   )
+  if (length(gamertag) == 1) return(resps[[1]])
+  make_collection(resps)
 }
 
 
@@ -44,7 +49,7 @@ halo_multiplayer <- function(
     .progress = TRUE
 ) {
   gamertag <- verify_vector_chr(gamertag, 'gamertag')
-  furrr::future_map(
+  resps <- furrr::future_map(
     .x = gamertag,
     .f = function(gamertag) {
       HaloDotAPI(
@@ -52,10 +57,13 @@ halo_multiplayer <- function(
         gamertag = gamertag,
         filter = null_or_chr(filter, 'filter'),
         version = version,
-        token = token
+        token = token,
+        class = 'multiplayer'
       )
     },
     .options = furrr::furrr_options(seed = NULL),
     .progress = .progress
   )
+  if (length(gamertag) == 1) return(resps[[1]])
+  make_collection(resps)
 }