+++ NEWS.md
@@ -1,5 +1,7 @@
 # halo (development version)
 
+* General improvements in S3 class handling
+
 * Added support for vectorized API requests
 
 * Added support for parallel processing of requests via the `{furrr}` package